Bring support for jax>=0.4.1 (#240)
Recently a new type `Array` was added to JAX to unify the different internal representations of arrays including `DeviceArray` (the one we were aliasing as `JaxArray`). This change causes `plum` to throw as it can not match the type of JAX arrays to the correct one in JAX versions >=0.4.1.

The solution here fixes such issue while maintaining compatibility with older JAX versions.