Upgrade HOOMD-blue and OpenMM in Colab (#341)
Since we've been supporting HOOMD-blue `v4.x` series for a while now, it makes sense to try to modernize our example notebooks. I'll start just with the Installation Notebook, and will update the examples separately.

This would in principle break our HOOMD-blue notebooks, but all notebooks are broken for other reasons, like #339. We will updated the notebooks in a separate PR.