[lammps] Fix handling types when restoring (#344)
For LAMMPS, only unique masses are stored, instead the mass "type" of each particle is stored, and the mass of the `i`th particle can be retrieved as `masses[types][i]`. For this reason we store `(masses, types)` in place of a `masses`, this requires special handling when restoring which we missed initially. Fixing it here.